# anotame-api/backend/sales-service/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    AddPaymentRequest:
      type: object
      required:
      - amount
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
...
    AuditLogResponse:
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UUID"
        fieldName:
          type: string
        oldValue:
          type:
          - string
          - "null"
        newValue:
          type:
          - string
          - "null"
        changedAt:
          $ref: "#/components/schemas/OffsetDateTime"
...
    CreateOrderRequest:
      required:
      - customer
      - items
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/CustomerDto"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDto"
        committedDeadline:
          type:
          - string
          - "null"
          anyOf:
          - $ref: "#/components/schemas/OffsetDateTime"
          - type: "null"
        notes:
          type:
          - string
          - "null"
        amountPaid:
          type:
          - number
          - "null"
        paymentMethod:
          type:
          - string
          - "null"
        priceListId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        priceListName:
          type:
          - string
          - "null"
...
    CustomerDto:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        firstName:
          type: string
          pattern: \S
        lastName:
          type:
          - string
          - "null"
        email:
          type:
          - string
          - "null"
        phoneNumber:
          type: string
          pattern: \S
        preferences:
          type:
          - object
          - "null"
          additionalProperties: {}
      required:
      - firstName
      - phoneNumber
...
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
    OrderContentSource:
      type: string
      enum:
      - CATALOG
      - CUSTOM
    OrderItemDto:
      type: object
      properties:
        garmentTypeId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        source:
          $ref: "#/components/schemas/OrderContentSource"
        garmentName:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemServiceDto"
        quantity:
          type: integer
          format: int32
        notes:
          type: string
      required:
      - source
    OrderItemResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        garmentTypeId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        source:
          $ref: "#/components/schemas/OrderContentSource"
        garmentName:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemServiceDto"
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type:
          - string
          - "null"
    OrderItemServiceDto:
      type: object
      required:
      - source
      - serviceName
      - unitPrice
      - durationMin
      properties:
        serviceId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        source:
          $ref: "#/components/schemas/OrderContentSource"
        serviceName:
          type: string
        unitPrice:
          type: number
        adjustmentAmount:
          type:
          - number
          - "null"
        adjustmentReason:
          type:
          - string
          - "null"
        durationMin:
          type: integer
          format: int32
        instructions:
          type:
          - string
          - "null"
    OrderResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        ticketNumber:
          type: string
        customer:
          $ref: "#/components/schemas/CustomerDto"
        committedDeadline:
          $ref: "#/components/schemas/OffsetDateTime"
        status:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        paymentMethod:
          type:
          - string
          - "null"
        notes:
          type:
          - string
          - "null"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        createdAt:
          $ref: "#/components/schemas/OffsetDateTime"
        totalDurationMin:
          type:
          - integer
          - "null"
          format: int32
        pickupCode:
          type:
          - string
          - "null"
        deliveredAt:
          type:
          - string
          - "null"
          anyOf:
          - $ref: "#/components/schemas/OffsetDateTime"
          - type: "null"
        priceListId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        priceListName:
          type:
          - string
          - "null"
    OrderSummaryPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    OrderSummaryResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        ticketNumber:
          type: string
        customer:
          $ref: "#/components/schemas/CustomerDto"
        committedDeadline:
          type:
          - string
          - "null"
          anyOf:
          - $ref: "#/components/schemas/OffsetDateTime"
          - type: "null"
        status:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        totalDurationMin:
          type:
          - integer
          - "null"
          format: int32
        createdAt:
          type:
          - string
          - "null"
          anyOf:
          - $ref: "#/components/schemas/OffsetDateTime"
          - type: "null"
        deliveredAt:
          type:
          - string
          - "null"
          anyOf:
          - $ref: "#/components/schemas/OffsetDateTime"
          - type: "null"
        garmentNames:
          type: array
          items:
            type: string
        serviceNames:
          type: array
          items:
            type: string
...
    PaymentResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        orderId:
          $ref: "#/components/schemas/UUID"
        amount:
          type: number
        paymentMethod:
          type:
          - string
          - "null"
        notes:
          type:
          - string
          - "null"
        recordedAt:
          $ref: "#/components/schemas/OffsetDateTime"
        orderAmountPaid:
          type: number
        orderTotalAmount:
          type: number
        orderBalance:
          type: number
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UpdateOrderRequest:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/CustomerDto"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDto"
        committedDeadline:
          $ref: "#/components/schemas/OffsetDateTime"
        notes:
          type: string
...
  securitySchemes:
    SecurityScheme:
      description: Authentication
      type: apiKey
      name: jwt
      in: cookie
paths:
...
  /orders:
    get:
      operationId: OrdersController_getOrders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Orders
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
    post:
      operationId: OrdersController_createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Order
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
...
  /orders/summary:
    get:
      operationId: OrdersController_getOrderSummaries
      parameters:
      - name: deadline
        in: query
        schema:
          $ref: "#/components/schemas/LocalDate"
      - name: garmentId
        in: query
        schema:
          $ref: "#/components/schemas/UUID"
      - name: garmentSource
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: search
        in: query
        schema:
          type: string
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: status
        in: query
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummaryPageResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Order Summaries
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
  /orders/{id}:
    put:
      operationId: OrdersController_updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update Order
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
    get:
      operationId: OrdersController_getOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Order
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
    delete:
      operationId: OrdersController_deleteOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete Order
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
  /orders/{id}/audit:
    get:
      operationId: OrdersController_getAuditLog
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Audit Log
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
  /orders/{id}/cancel:
    patch:
      operationId: OrdersController_cancelOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Cancel Order
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
...
  /orders/{id}/status:
    patch:
      operationId: OrdersController_updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Update Status
      tags:
      - Orders Controller
      security:
      - SecurityScheme: []
  /orders/{orderId}/payments:
    get:
      operationId: OrderPaymentController_getPayments
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Payments
      tags:
      - Order Payment Controller
      security:
      - SecurityScheme: []
    post:
      operationId: OrderPaymentController_addPayment
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPaymentRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
          description: Created
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Add Payment
      tags:
      - Order Payment Controller
      security:
      - SecurityScheme: []
...
info:
  title: sales-service API
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8083
  description: Auto generated value
- url: http://0.0.0.0:8083
  description: Auto generated value